chore: eliminate require cycle warning for intl-related code (#342)
